test: Set context timeout after setup stage
As now run of the praefect service requires a database
creation and this process is synced with other parallel
operations doing the same we need to address a timeout
set of the context for execution. The setup code is not
part of the actual verification that is why timeout set
on the context only after database setup is done.