Link repository id to replicas created from replication jobs
SetGeneration is called to set a replica's generation following a replication
job. As it might create new replica records, it should ensure the record is
linked to the correct repository via the repository ID. This commit links the
replica's record by getting the correct repository ID from the repositories table.
As all replicas still have the same (virtual_storage, relative_path) as the
repository, this is still safe to do. It's also necessary as not every replication
job yet carries a repository ID. This commit also includes the repository ID in
replication jobs so the approach can later be updated to simply take the repository
ID from the replication job.