smarthttp: Verify repo existence before dealing with cache
InfoRefsUploadPack and InfoRefsReceivePack doesn't verify
if repository exist before the caching get involved into the
process. It requires additional operations to run that does
nothing at the end because of missed repository. It also
produces error log entries that may confuse some time.
In order to optimise this case we first get path to the
repository and if it doesn't exist the error is returned
immediately without need to bother caching (and no extra
logs).