operations: Support object quarantine in UserRebaseConfirmable
While almost all RPCs which do access checks have been converted to use
quarantine directories, `UserRebaseConfirmable()` was an oversight
that's not yet been converted. Fix this by using `quarantinedRepo()`
instead of `localrepo()` and passing the quarantine directory into
`updateWithHooks()`. Like this, if the "Quarantine" feature flag is set,
we'll create a quarantined repository and only migrate new objects in
case hooks accept the change.

Note that the "Quarantine" feature flag is a preexisting feature flag.
Reusing it should be fine though given that the implementation of all
parts has already been tested and is not expected to cause new issues.
If it does raise issues, we can still disable the feature flag and thus
disable quarantine directories altogether via this single flag instead
of growing additional quarantine-related flags.

Changelog: performance