testhelper: Rename `ConfigureGitalyLfsSmudge()`
Rename `ConfigureGitalyLfsSmudge()` to `BuildGitalyLFSSmudge()`: this
highlights more clearly that the intention of the function is to build a
command in contrast to configuring the command.

Furthermore, adjust the function to have the same signature and call
path as the other binary-building helpers. While the overhead of
`buildBinary()`, which also makes sure that the binary is built exactly
once, is not really needed considering we only build the command once in
all of our tests, unifying the code path will aid us to refactor the
build helper.