testhelper: Fix timeouts when building executables
Our tests rely on several executables to exist in the binary path, which
is why we have a set of helpers which builds these executables for us.
Given that the build may take some time, which quickly adds up if done
dozens of times, we have locking in place which avoids rebuilds of the
binary in case we've already built it. This locking is currently done
via a lockfile. Once a process acquires the lockfile, the others then
poll it until it's removed and the binary has been successfully built.

With our recent move to use `t.Parallel()` in more places, the polling
has been seen to time out much more frequently. This can probably be
explained with increased load, where building the binary now takes
longer than the 5 seconds deadline we give it.

Refactor the locking code to not use lockfiles, but instead `sync.Once`:
this kind of code is exactly what it's designed for, where we want to
set up code once. With the new and simplified code, we also drop any
deadlines for building the code: we don't want to exercise how fast the
compiler is, but only care for the fact that it finishes eventually. If
it started to fail at some point in time, then we sure would notice fast
enough even without having a timeout.

Given that we don't need polling anymore with `sync.Once`, the new code
is also slightly faster with a lot less variance:

    Benchmark #1: lockfile
      Time (mean ± σ):     35.679 s ±  0.651 s    [User: 290.002 s, System: 81.925 s]
      Range (min … max):   34.964 s … 36.239 s    3 runs

    Benchmark #2: sync.Once
      Time (mean ± σ):     35.121 s ±  0.093 s    [User: 290.491 s, System: 81.848 s]
      Range (min … max):   35.015 s … 35.191 s    3 runs

    Summary
      'sync.Once' ran
        1.02 ± 0.02 times faster than 'lockfile'

So it's a win-win: the code is both simpler and faster.