PostUploadPack: split partial clone test
When partial clone support was introduced in Gitaly, it was disabled
by default, and it required setting allowAnySha1InWant on a
per-request basis. So the test for partial clone tested both partial
clone and allowAnySha1InWant. Now allowAnySha1InWant is on globally,
so it makes no sense to regard it as part of partial clone anymore.
This commit splits allowAnySha1InWant into its own test, which is
better because now its clearer what each test is about.