Add method for getting a repository's ID
Praefect receives requests from the client to repositories identified
by a `(virtual_storage, relative_path)` tuple. Praefect needs to lookup
the correct repository ID for this tuple at the time of the request.
Doing the lookup once for each request and operating on the repository
id afterwards guarantees that any in-flight jobs and requests will always
target the same repository. If Praefect depended on the tuple, concurrent
modification could cause the updates to target wrong repositories in the
database, for example if a repository is renamed while a request is in
flight and generation increment is attempted with the old name after the
request. This is not really a problem in practice though but is fixed by
this nonetheless. In practice, Praefect needs the GetRepositoryID to propagate
the repository's ID in any replication jobs a mutator creates. This ensures any
new replicas created from the replication jobs get linked to the correct
repository.