Add PackObjectsHookWithSidechannel RPC
This RPC is a replacement for PackObjectsHook. It uses a unix socket
sidechannel for data transfer, which is worthwhile because
PackObjectsHook responses can be very large (gigabytes), and
transferring gigabtes of data over a gRPC stream has a high overhead.

This is part of https://gitlab.com/groups/gitlab-com/gl-infra/-/epics/463.