helper: Implement helper to create gRPC errors with details
The gRPC error model allows for two different usages: first the standard
error model where one can return a simple error message associated with
a code. And then the richer error model, where structured data can be
added to errors such that details about the failure mode can passed to
the caller [1]. In Gitaly, we always use the simple error model.

This simple error model is quite limiting though. For example in the
operations service, we have resorted to pass error information to the
caller by embedding error details into the normal response. While not
elegant, it has served us alright in the past. But this design is
awkward, has issues with Praefect if it inspects proxied errors, and is
just bad design in general.

To fix this, this commit implements a new helper function which will
enrich a gRPC status error with additional details. These details can be
an arbitrary protobuf message. Like this, we can pass up arbitrary
structured data to the caller, which will allow us to fix these
shortcomings.

[1]: https://www.grpc.io/docs/guides/error/