hook: Fix prereceive returning AllowedError for generic errors
The pre-receive hook implementation is responsible for calling GitLab's
`/internal/allowed` API to do access checks. In case this API endpoint
tells us that a change is not allowed, we need to make the error message
available to the original caller, which then knows to return the error
message via the response's `PreReceiveError` field.

Our current implementation of this is wrong though: while we correctly
pass up information via a `NotAllowedError` in case the access check
indicates that the change is not allowed, we also pass up information
when invoking the endpoint itself failed. But this really is an internal
error, and we don't ever want to return such information to the user.

Fix this issue by returning a normal error in such cases.

Changelog: fixed