gitaly-git2go: Support skipping empty rebased commits
In case the diff a commit introduces has already been applied upstream
the commit will become an empty commit on rebase. In such cases, libgit2
will return an error to alert the developer. And while this is something
that we want (for now) in normal rebases, it does break squashing of a
range of commits in case this range contains now-empty commits.

Add a new option to the rebase command which allows the caller to skip
any such empty commits without an error. This doesn't yet change
behaviour given that there are no callers yet which set this new option.