gitaly-git2go: Fix rebase containing wrong commit if rebase is empty
When rebasing a range of commits, then it can happen that the rebase
itself does not contain any commits. This can either happen in case
when the end commit is already contained in the start commit, or
alternatively if all changes in the range between merge base and end
commit have already been applied to the start commit and thus become
empty. In this case, the rebased commit is equal to the start commit
onto which we have rebased: rebasing nothing onto a start commit will
never yield any changed commit. But what we return from the rebase
commit is the end commit.

Fix this by correctly returning the start commit in this case. This is
only a theoretical fix for now: we detect early whether the end commit
is already contained in the start commit, and we return an error in case
any of the rebased commits starts to become empty. The latter is about
to change though given that we'll allow rebasing of emtpy commits, so
this fix prepares for that future.