Remove redundant struct name in RepositoryStore tests
RepositoryStore tests use helper types to represent state to be
asserted in the tests. As the map type already contains the type of the
struct it contains, it's not necessary to write out the struct name when
writing a literal map. This commit removes the redundant struct name from
the literals.