test: Parallel tests dependent on the PostgreSQL database
As now we create a new database on each call of the
glsql.GetDB the test execution time increased. To mitigate
that problem we make tests dependent on the PostgreSQL database
parallel. It should reduce test execution time and may help us
to find new issues in the code because of the shared global state.