Do not inject replication queue into InfoService
InfoService previously used the replication queue to schedule replication
jobs as part of the manual reconciliation command and when accepting
data loss. The manual reconciler has been removed and accepting data loss
relies on the automatic reconciler. As the InfoService no longer needs to
queue replication jobs, this commit removes the unnecessary field and updates
the call sites not to inject the queue.