ref: Always enable gitpipe-based FindAllTags implementation
The FindAllTags implementation based on the gitpipe packages has been
tested in production now for some time without any additional problems.
While it hasn't proven to be the performance improvement I was hoping
for (it performs roughly the same), the new implementation shares more
code with other implementations.

Remove the feature flag and always enable the gitpipe implementation.