Unify backup restore skipping
Restores were already skipped when the bundle did not exist. This
extends the logic further to say that restores will be skipped if the
locator returns `ErrDoesntExist`.

The returned errors for the filesystem sink were less descriptive than
for the storage service sink. So these errors were updated to make them
equivalent.