repository: Deprecate `SetConfig()` RPC
With the `SetConfig()` RPC, we used to allow Rails to modify the on-disk
config of a Git repository. As a result, we have on-disk state that is
hard to observe for Gitaly and which can lead to hard-to-debug issues
when the config contains unexpected entries. We have thus worked towards
the goal of having complete control over the gitconfig in Gitaly:

    - Users of on-disk remotes now use in-memory ones.

    - JWT tokens are passed via header fields in requests which fetch
      objects from a remote repository, where the token is injected via
      environment variables.

    - The last valid usecase of modifying the gitconfig is to set
      "gitlab.fullpath". This usecase has since grown its own RPC call
      `SetFullPath()`.

As a result, we don't have any users left which call either
`SetConfig()` or `DeleteConfig()`. This commit thus marks both as
deprecated so we can remove them in the next release.

Changelog: deprecation