Return a better error message when a repository has no default branch
UserApplyPatch applies the patches it receives on the default branch.
The Ruby implementation that is now disabled returned a TypeError when
attempting to operate on a nil object that is returned when the repository
has no default branch. The Go port has a more descriptive error for the
case but is currently matching the Ruby implementation's error. This
commit returns the more descriptive error now that the Ruby implemntaion
is disabled.

Changelog: fixed