updateref: Generalize `PreReceiveError` according to its usage
When any of the hooks fails to execute in our `UpdaterWithHooks`, then
we return a `PreReceiveError` to the caller. This is quite misleading
though: we not only execute the pre-receive hook, but also others, and
they all return the same error.

Refactor the error to instead be a `HookError`. While at it, we also
embed stdout and stderr in the error such that it can handle generation
of the error message itself instead of using `hookErrorMessage()`, and
implement `Unwrap()` such that we can use `errors.As()` to retrieve the
embedded error. This will be used at a later point to distinguish
`HookError`s and `NotAllowedError`s.