testserver: Build Praefect executable ad-hoc
In order to execute tests with Praefect, we currently build Praefect as
a test prerequisite. There is not much of a point in doing so, but we
can instead treat it the same as we treat all the other executables
which we rely on during tests and just build it ad-hoc.

Refactor the code to do so.