coordinator: Fix modification of envvars to change testing behaviour
In our tests, we typically only create a repository on-disk without also
creating the database entries in case we're running with Praefect as
proxy. As a result, all repository-scoped RPC calls via Praefect would
fail because it doesn't know the repository. As a workaround, we're thus
listening on the GITALY_TEST_PRAEFECT_BIN environment variable: if it's
set, then we force-create repository entries in the database such that
we don't run into this issue.

In some tests, we _do_ want to test the real logic though, which is why
we unset the envvar in these tests. But modification of the environment
is always for the complete process, and combined with out recent move to
use `t.Parallel()` this can have unintended sideeffects for other tests
which run concurrently.

Fix this issue by introducing a new config variable for Praefect which
allows us to toggle this behaviour. This easily allows us to control
behaviour both for spawned Praefect executables, where we write an
ad-hoc config, and for in-memory instances where we pass in a Praefect
config struct.

The variable obviously shouldn't ever be used in production, but based
on it's naming this is quite clear.