testhelper: Do not copy binaries in case they already exist
When building binaries, we always copy the resulting binaries into the
target location. This is not required though in case we have already
built and copied the binary before, and in fact it even causes issues if
the binary exists already and is still in use.

Fix this by not copying the binary in case it already exists.