featureflag: Add constructor to avoid having to modify `All`
The `All` array of feature flags is required for our tests such that we
can easily toggle on all feature flags regardless of their default
value. Keeping track of all feature flags like this is error prone
though: it frequently happens that feature flags are not added to the
array, which decreases our test coverage. Furthermore, it often leads to
conflicts when another feature flag has been modified meanwhile.

Fix this by adding a new constructor for feature flags: next to creating
the feature flag itself, it will also add the feature flag to the array.