gitaly-git2go: Split merge code into runner and actual logic
The gitaly-git2go merge code handles deserialization of params,
deserialization of results and the actual merge logic in a single
function. This makes it hard to reason about inputs and outputs of this
function and thus prohibits an easy refactor of them.

Refactor the code and pull out the merge logic into its own function,
where all parameters are obtained as input and output is handled by the
calling function.

No change in behaviour is expected from this refactoring.