gitaly-git2go: Fix merge response always going to stdout
Originally, gitaly-git2go commands would always write to their responses
to stdout themselves. With the introduction of gob for serialization and
deserialization of inputs and outputs, this has changed to pass in a
reader and writer to all commands. The merge subcommand is still
directly writing to stdout though.

Fix this issue by writing to the passed-in writer instead. This is only
a theoretical issue: there are no callers which would ask gitaly-git2go
subcommands to write anywhere else.