gitaly-git2go: Support gob in the merge subcommand
Back when we have started implementing RPCs via gitaly-git2go, the
initial implementations were all using JSON as representation for inputs
and outputs. At a later point we noticed that this makes it hard to
return typed errors, which is why later subcommands have shifted to
using gob encoders.

The merge subcommand is the first command we have written, and thus it
also uses JSON encoding, facing the exact error handling issue mentioned
above. We want to extend the command to return information about
conflicting files though, which would ideally be part of the returned
error type.

Refactor the subcommand to optionally support gob as encoding: if the
command is called with no JSON request, then we instead use gob. While
this seems kind of cheap, it's easier than adding a separate flag and
should always provide expected results: we never call the command
without a request, and the request cannot ever be empty given that it's
a JSON-encoded struct.