git2go: Allow registration of gob types with slices
In order to use gob for the serialization of types, we first need to
register types with gob. This is done in our internal git2go package,
where we loop through all types in a map, calling `gob.Register()` for
each of them. Later on, when we are trying to serialize errors, we'll
see whether the error unwraps to any of these registered types by
looking up the zero-value of the current error in that map.

This has an important shortcoming though: maps can only hold values
which are comparable. While most of the standard types are, slices are
not and thus it's impossible to put a slice or a struct containing a
slice into a map. We thus cannot currently register any types which make
use of slices.

Fix this shortcoming by tracking `reflect.Type`s in the map instead of
zero value.