PackObjectsHookWithSidechannel: return Canceled when client hangs up
PackObjectsHookWithSidechannel replaces PackObjectsHook. We used to
regularly see a low number of PackObjectsHook fail with the 'Canceled'
status code. Now PackObjectsHookWithSidechannel fails at a similar low
rate with 'Unknown'. With gRPC calls, we have a server middleware that
sets the status code to Canceled in cases where the client hung up.
This middleware does not work for sidechannel connections so that
explains the 'Unknown' codes.

This change fixes the error code returned by
PackObjectsHookWithSidechannel in the case that the client hangs up
while using the side channel.

Changelog: fixed