Makefile: Fix Go sources not including protoc-gen-gitaly plugin
When executing the formatter, we use `find_go_sources` to find the
set of Go files which should be formatted. This function excludes a
bunch of files which shouldn't ever be formatted, like for example
generated protobuf files. These exclusions have been too broad though
and caused us to never format sources of our protoc-gen-gitaly command,
which is also contained in "proto/go/".

Fix this by only excluding "proto/go/gitalypb/", which contains all
generated sources, and run the formatter on the newly found sources.