safe: Add option to change file mode for FileWriter
When using the FileWriter interface, then the resulting file's mode will
always be set according to the current umask, and there aren't currently
any callers which would need to change that mode. We're about to convert
callers to use this interface though, and they have specific needs with
regards to the resulting file mode.

While it would be possible for callers to manually set the file mode,
this would be error prone due to the safety guarantees we want to make
with this interface. Most importantly, we want to make sure that changes
have been written to disk before committing, which may easily be
forgotten by callers.

Adds a new optional config parameter to `NewFileWriter()` which allows
callers to specify a mode to implement this upcoming use case.