safe: Implement new locking file writer
When updating files in a transaction, then we want to make sure that
there are no two concurrent callers which can somehow interleave voting
on and writing to the file, lest the end result is unclear. Doing this
correctly is tricky though: while we could simply take a lockfile during
the whole operation and then modify the file in-place, it's possible
that the critical section takes a long time and meanwhile blocks all
other writers. Alternatively, concurrent readers may see in-between
state when the file is being updated in place.

Introduce a new locking file writer which encapsulates this logic. This
new interface is based on the FileWriter interface: given a target file,
it will first create a temporary file next to that target file and then
atomically rename the file into place when all updates have been written
to the temporary file. The difference though is that in order to commit
the file, the caller first needs to lock the target file such that no
other user can concurrently modify it at the same time. Furthermore, to
avoid overwriting data written by a concurrent writer, we will ensure
that the target file has not been modified between creating the writer
and locking it.

With this interface, we can ensure that the update is atomic while the
critical section is as short as possible.