Only use HEAD to determine default branch
The original default branch logic would select the only branch that
existed, or arbitrarily choose the first branch if it couldn't find a
better option. In addition it was dereferencing to the commit even
though no callees were using the object ID.

Here we're changing this logic to more heavily rely on HEAD and only
ever falling back to default branches if they exist, without
dereferencing to the commit.

Changelog: changed