Make having no default branch an error in GetDefaultBranch
Instead of returning an invalid empty ReferenceName we return an error
indicating that no default branch could be found. This forces the caller
of GetDefaultBranch to decide if it matters in their specific case.

For example, ref service previously filtered for merged branches without
first checking if there was actually a default branch even though the
filter should have been impossible without it.