repoclean: Introduction of the new storage_cleanups table
The problem with repositories missing in the praefect database
becomes more actual as more customers migrate to the cluster
setup. In order to verify state of the things in the cluster
the new background job needs to be implemented. As the first
step it should run over all repositories on gitaly storages
and check repository existence in the praefect database.
As usually there are multiple instances of the praefect in the
cluster to support HA we need to coordinate scan by each
praefect. That is why a new table storage_cleanups is introduced.
It is used to coordinate available for scan gitaly storages.
The coordination is based on the acquiring and "locking" (write
to the column) of the rows, so acquired rows are not visible
for other praefect instances until execution is in progress.

The change also includes method to verify existence of the
repositories in the praefect database.

Part of: https://gitlab.com/gitlab-org/gitaly/-/issues/3719