transaction: Implement function to easily vote and commit locked files
Initially, transactional voting has only been implemented when updating
Git references. Given that these use the reference-transaction hook, we
know that we have proper locking semantics because Git will first lock
the references which are about to be updated, then call the hook in
"prepared" state, commit locked references to disk and then call the
hook in "committed" state again. That is, at the point of our first vote
via this hook, all references are locked and cannot be changed by any
other concurrent caller.

We have since extended this mechanism to non-reference files, like
gitattributes or gitconfig. In the current implementation, we do not
provide the same safeguards though as described above: there is no
locking, no checking for concurrent writes and no guaranteed atomicity.

This is about to change via the new LockingFileWriter interface. For any
given file update, we now want to do the following sequence in order to
commit file changes:

    1. Create a new LockingFileWriter for the target file, optionally
       containing the same content as the current target file.

    2. Update the temporary file to contain the desired changes.

    3. Lock the file such that no concurrent changes are possible to the
       target file anymore. At this point in time, we also verify that
       the target file hasn't been updated while we were updating our
       copy of it. This step is required to make sure that there was no
       race with another process.

    4. Vote on the updated file contents. This is required to assert
       that the Gitaly cluster performed the same change on each node.

    5. Commit the file by moving it into place. The change is now
       visible to other clients and the lock has been released.

    6. Do another vote to demonstrate that the modification has been
       performed.

This sequence is rather involved, and it is easy to mix up any of the
steps. Implement a new function `transaction.CommitLockedFile()` which
encapsulates the logic such that we only need a single implementation of
it.