Link existing database record via repository ID
Praefect generates a repository ID to uniquely identify a repository.
While new records that get created are already being linked via the
repository ID, historical records in the database are still not linked
via the repository ID. In order to update the queries to join the records
via the repository ID, all of the relevant records need to be linked via
the repository ID. This commit adds a migration that links the records
in 'storage_repositories' and 'repository_assignments' to the 'repositories'
table's records via the ID.

Replication jobs are not included in the migration as they are not long lived
and they can be rescheduled.

Changelog: other