coordinator: Fix test races caused by concurrent repository creation
Because most of our tests won't ever create repository records in the
database when they set up a repository, we have a workaround in the
coordinator which creates database records for repositories whenever we
see a repository-scoped message. This approach is currently racy though:
in case we have multiple concurrent repository-scoped RPC calls, we'll
try to create the same repository multiple times. The result is that one
of both RPCs will fail with an error.

Fix this issue by ignoring `RepositoryExistsError`s when force-creating
the database entries.