Use Postgres in RouteRepositoryMutator tests
RouteRepositoryMutator do not use Postgres currently and thus do not
test as much of the production code as possible. This commit replaces
the mocks in the tests with real implementations using Postgres.

Some of the test cases had to be adjusted due to the lazy election
logic in Praefect which is now promoting a new primary instead
of failing the RPC if there is a primary candidate available.