Use repository IDs in IncrementGeneration
IncrementGeneration currently identifies the records to modify by the
virtual storage and relative path tuple. This needs to be changed as
the replicas will soon have Praefect generated relative paths which don't
match the relative path the client uses to access the repository. Relying
on the external relative path is also prone to cause races as it or the
repository it refers to could change during an in-flight request. Relying
on the repository ID retrieved at the beginning of the request handling
ensures all database operations always refer to the correct repository.